+++ CHANGELOG.md
@@ -42,3 +42,4 @@
 - Date inputs in forms for creating activity, transaction and a budget have a hint text
 - Transaction provider and receiver IATI references are exposed in the XML if present
 - Users can report project level activities
+- Users can add budgets to project level activities